Moved local_log_sum_exp to special.py
